requirements.txt
annotated-types==0.7.0
anyio==4.8.0
atproto==0.0.58
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.8
fonttools==4.55.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libipld==3.0.1
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
psycopg2-binary
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-dotenv>=1.0.0
pytrends==4.9.2
pytz==2025.1
requests==2.32.3
rich==13.9.4
seaborn==0.13.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
thinc==8.3.4
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
websockets==13.1
wrapt==1.17.2